Add BazelActionQuery and BSPError modules; refactor BuildServer and RequestHandler
